requirements.txt
fastapi==0.104.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.9
pydantic==2.3.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.9.1
webdriver-manager==3.8.6
requests==2.31.0
aiohttp==3.8.5
pydantic-settings==2.1.0
streamlit==1.32.0
alembic==1.12.0